removed redundant mecadrive
